+++ src/cpp/core/StringUtils.cpp
@@ -799,6 +799,44 @@ std::string formatDouble(const double d, const int precision)
    return out.str();
 }
 
+std::string sprintf(const char* fmt, ...)
+{
+   // note: the semantics for vsnprintf are slightly awkward... when vsnprintf
+   // is called with a null pointer, it returns the number of characters that
+   // would be written, not including the null terminator. however, when called
+   // with a buffer, vsnprintf will write a maximum of n - 1 characters, and
+   // will always write a null terminator at the end! so we need to ensure we
+   // add 1 character to the size returned by vsnprintf(nullptr) to get the
+   // full size of the C string we want to generate
+   std::size_t n = 0;
+   {
+      va_list args;
+      va_start(args, fmt);
+      n = std::vsnprintf(nullptr, 0, fmt, args);
+      va_end(args);
+   }
+   
+   if (n == 0)
+   {
+      return std::string();
+   }
+   
+   // allocate buffer of required size
+   // (include space for null pointer)
+   std::vector<char> buffer(n + 1);
+   
+   // write formatted string to buffer
+   {
+      va_list args;
+      va_start(args, fmt);
+      std::vsnprintf(&buffer[0], buffer.size(), fmt, args);
+      va_end(args);
+   }
+   
+   // return as string
+   return std::string(&buffer[0], n);
+}
+
 } // namespace string_utils
 } // namespace core 
 } // namespace rstudio